add simple regression test and move the test to a gist: https://gist.github.com/mattbennett/1836aa08c4e640269e36ad6986df66a8
