+++ test/test_queue_consumer.py
@@ -1,4 +1,6 @@
+import itertools
 import socket
+import time
 import uuid
 
 import eventlet
@@ -7,11 +9,15 @@ from eventlet.event import Event
 from kombu import Connection, Exchange, Queue
 from kombu.exceptions import TimeoutError
 from mock import ANY, Mock, call, patch
+
 from nameko.constants import (
     AMQP_URI_CONFIG_KEY, DEFAULT_HEARTBEAT, HEARTBEAT_CONFIG_KEY)
+from nameko.events import event_handler
 from nameko.messaging import QueueConsumer
-from nameko.rpc import rpc
+from nameko.rpc import RpcProxy, rpc
+from nameko.standalone.events import event_dispatcher
 from nameko.standalone.rpc import ServiceRpcProxy
+from nameko.testing.services import entrypoint_waiter
 from nameko.testing.utils import (
     assert_stops_raising, get_extension, get_rabbit_connections)
 
@@ -378,6 +384,79 @@ class TestHeartbeats(object):
 
         queue_consumer = get_extension(container, QueueConsumer)
         assert queue_consumer.connection.heartbeat == heartbeat
+
+
+class TestDeadlockRegression(object):
+    """ Regression test for https://github.com/nameko/nameko/issues/428
+    """
+
+    @pytest.fixture
+    def config(self, rabbit_config):
+        config = rabbit_config.copy()
+        config['max_workers'] = 2
+        return config
+
+    @pytest.fixture
+    def upstream(self, container_factory, config):
+
+        class Service(object):
+            name = "upstream"
+
+            @rpc
+            def method(self):
+                time.sleep(.5)
+
+        container = container_factory(Service, config)
+        container.start()
+
+    @pytest.fixture
+    def service_cls(self):
+
+        class Service(object):
+            name = "downsteam"
+
+            upstream_rpc = RpcProxy("upstream")
+
+            @event_handler('service', 'event1')
+            def handle_event1(self, event_data):
+                self.upstream_rpc.method()
+
+            @event_handler('service', 'event2')
+            def handle_event2(self, event_data):
+                self.upstream_rpc.method()
+
+        return Service
+
+    @pytest.mark.usefixtures('upstream')
+    def test_deadlock_due_to_slow_workers(
+        self, service_cls, container_factory, config
+    ):
+        """ Deadlock will occur if the unack'd messages grows beyond the
+        size of the worker pool at any point. The QueueConsumer will block
+        waiting for a worker and pending RPC replies will not be ack'd.
+        Any running workers therefore never complete, and the worker pool
+        remains exhausted.
+        """
+        container = container_factory(service_cls, config)
+        container.start()
+
+        count = 2
+
+        dispatch = event_dispatcher(config)
+        for _ in range(count):
+            dispatch("service", "event1", 1)
+            dispatch("service", "event2", 1)
+
+        counter = itertools.count(start=1)
+
+        def cb(worker_ctx, res, exc_info):
+            if next(counter) == count:
+                return True
+
+        with entrypoint_waiter(
+            container, 'handle_event1', timeout=5, callback=cb
+        ):
+            pass
 
 
 def test_greenthread_raise_in_kill(container_factory, rabbit_config, logger):
